+++ badge/src/AdapterForReadingBadgeContextValues/PackagistContextValueReader.php
@@ -2,7 +2,7 @@
 
 namespace Badge\AdapterForReadingBadgeContextValues;
 
-use Badge\Application\PortOut\ContextValueReader;
+use Badge\Application\Port\Driven\ForReadingBadgeContextValues\ContextValueReader;
 use Badge\Infrastructure\ReleasedVersions;
 use Packagist\Api\Client;
 use Packagist\Api\Result\Package;
