+++ badge/src/Application/Port/Driven/ForReadingBadgeContextValues/ContextValueReader.php
@@ -1,0 +1,15 @@
+<?php declare(strict_types=1);
+
+namespace Badge\Application\Port\Driven\ForReadingBadgeContextValues;
+
+interface ContextValueReader extends
+    ForReadingSuggesters,
+    ForReadingDependents,
+    ForReadingTotalDownloads,
+    ForReadingMonthlyDownloads,
+    ForReadingDailyDownloads,
+    ForReadingStableVersion,
+    ForReadingUnstableVersion,
+    ForReadingLicense
+{
+}
